+++ leggedsnake/physicsengine.py
@@ -1,12 +1,16 @@
 # -*- coding: utf-8 -*-
 """
+This module gives a dynamic behavior to legged mechanism.
+
+Is uses the 2D physics engine chipmunk, this is why it can only be used on
+planar mechanisms.
+In theory you can use any type of mechanism, and not only planar mechanisms.
+In practice we do generate the road and some other parameters as the gravity,
+so it can be difficult to test something other than a walker.
+
 Created on Sat May 25 14:56:01 2019.
 
 @author: HugoFara
-
-This module gives dynamic behavior of legged mecanism.
-Is uses the 2D physics engine chipmunk, and can only be used on planar
-mecanisms.
 """
 import numpy as np
 import pymunk as pm