+++ sphinx/source/conf.py
@@ -22,7 +22,7 @@ copyright = '2021, Hugo Farajallah'
 author = 'Hugo Farajallah'
 
 # The full version, including alpha/beta/rc tags
-release = '0.1.2-alpha'
+release = '0.1.3-alpha'
 
 
 # -- General configuration ---------------------------------------------------
@@ -31,8 +31,6 @@ release = '0.1.2-alpha'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    # For Markdown support
-    #'myst_parser',
     # Use docstrings
     'sphinx.ext.autodoc',
     # Useful for markdown integration
@@ -46,7 +44,7 @@ templates_path = ['_templates']
 # Files to be used as source
 ource_suffix = {
     '.rst': 'restructuredtext',
-    '.md': 'markdown', # incompatible with m2r2
+    '.md': 'markdown',
 }
 
 # List of patterns, relative to source directory, that match files and
@@ -60,9 +58,9 @@ exclude_patterns = ['docs', 'Thumbs.db', '.DS_Store']
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'alabaster'
+html_theme = 'sphinx_rtd_theme'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['../../docs/html/_static']
+html_static_path = ['../../docs/_static']